requirements.txt
discord.py
python-dotenv
gTTS
PyNaCl
asyncpg
aiohttp
Pillow
pandas
tabulate
selenium
beautifulsoup4
numpy
psycopg2-binary
pdf2image
matplotlib
japanize-matplotlib
gspread
oauth2client
gspread-dataframe
jaconv